requirements.txt
aiogram==3.13
aiohttp
asyncpg
chapa
python-dotenv
openai
scipy
